Implement LevelList with updated data structures
