+++ Assets/Data/ScriptableObjects/LevelList.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9ef5a72804184dc4faee2ab81fe172cf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
